backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.0.3
aiofiles==23.2.1
Pillow==10.1.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1

# 文档解析库
pandas==2.1.4
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
markdown==3.5.1
pdfplumber==0.10.3
# python-pptx==0.6.23  # PowerPoint支持（可选）